requirements.txt
PyYAML==3.10
python-mpd==0.3.0
